requirements.txt
openai>=1.0.0
aiortc>=1.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
